requirements.txt
# Python requests library
requests==requests==2.18.1

# utilize 12-factor inspired environment variables
django-environ==0.4.3
